project2/requirements.txt
pandas
openpyxl
vanna
psycopg2-binary
google-generativeai
python-dotenv
google-cloud-aiplatform
